Move VolCreateRequest to where it should belong
