+++ volume/struct.go
@@ -33,9 +33,7 @@ var (
 	// ValidateBrickEntriesFunc validates the brick list
 	ValidateBrickEntriesFunc   = ValidateBrickEntries
 	validateBrickPathStatsFunc = utils.ValidateBrickPathStats
-	// NewVolinfoFunc returns an empty Volinfo
-	NewVolinfoFunc = NewVolinfo
-	absFilePath    = filepath.Abs
+	absFilePath                = filepath.Abs
 	// NewBrickEntriesFunc creates the brick list
 	NewBrickEntriesFunc = NewBrickEntries
 )
@@ -97,53 +95,6 @@ type VolStatus struct {
 	Brickstatuses []brick.Brickstatus
 	// TODO: Add further fields like memory usage, brick filesystem, fd consumed,
 	// clients connected etc.
-}
-
-// VolCreateRequest defines the parameters for creating a volume in the volume-create command
-// TODO: This should probably be moved out of here.
-type VolCreateRequest struct {
-	Name            string   `json:"name"`
-	Transport       string   `json:"transport,omitempty"`
-	DistCount       int      `json:"distcount,omitempty"`
-	ReplicaCount    int      `json:"replica,omitempty"`
-	StripeCount     int      `json:"stripecount,omitempty"`
-	DisperseCount   int      `json:"dispersecount,omitempty"`
-	RedundancyCount int      `json:"redundancycount,omitempty"`
-	Bricks          []string `json:"bricks"`
-	Force           bool     `json:"force,omitempty"`
-}
-
-// NewVolinfo returns an empty Volinfo
-func NewVolinfo() *Volinfo {
-	v := new(Volinfo)
-	v.Options = make(map[string]string)
-
-	return v
-}
-
-// NewVolumeEntry returns an initialized Volinfo using the given parameters
-func NewVolumeEntry(req *VolCreateRequest) (*Volinfo, error) {
-	v := NewVolinfoFunc()
-	if v == nil {
-		return nil, errors.ErrVolCreateFail
-	}
-	v.ID = uuid.NewRandom()
-	v.Name = req.Name
-	if len(req.Transport) > 0 {
-		v.Transport = req.Transport
-	} else {
-		v.Transport = "tcp"
-	}
-	if req.ReplicaCount == 0 {
-		v.ReplicaCount = 1
-	} else {
-		v.ReplicaCount = req.ReplicaCount
-	}
-	v.StripeCount = req.StripeCount
-	v.DisperseCount = req.DisperseCount
-	v.RedundancyCount = req.RedundancyCount
-
-	return v, nil
 }
 
 // NewBrickEntries creates the brick list
